+++ gamest_plugins/process_identifier/__init__.py
@@ -1,0 +1,3 @@
+from .module import ProcessIdentifierPlugin
+
+plugin = ProcessIdentifierPlugin